+++ CMakeLists.txt
@@ -35,7 +35,7 @@ endif()
 
 # Set install locations
 set( DD_BINARY_INSTALL_LOCATION "${CMAKE_INSTALL_PREFIX}/tools" )
-set( DD_INCLUDE_INSTALL_LOCATION "${CMAKE_INSTALL_PREFIX}/include" )
+set( DD_INCLUDE_INSTALL_LOCATION "${CMAKE_INSTALL_PREFIX}/include/double-down" )
 set( DD_LIBRARY_INSTALL_LOCATION "${CMAKE_INSTALL_PREFIX}/lib" )
 set( DD_CMAKE_CONFIG_LOCATION "${DD_LIBRARY_INSTALL_LOCATION}/cmake")
 
@@ -52,6 +52,9 @@ set(CMAKE_INSTALL_RPATH_USE_LINK_PATH TRUE)
 
 # add library
 add_library( dd SHARED "src/RTI.cpp" "src/primitives.cpp" "src/MOABRay.cpp" "src/MOABDirectAccess.cpp")
+target_include_directories(dd PRIVATE ${CMAKE_SOURCE_DIR}/include)
+set_target_properties(dd PROPERTIES INSTALL_RPATH_USE_LINK_PATH TRUE)
+set_target_properties(dd PROPERTIES INSTALL_RPATH ${CMAKE_INSTALL_PREFIX}/lib )
 
 target_include_directories( dd INTERFACE "${CMAKE_INSTALL_PREFIX}/include" ${EMBREE_INCLUDE_DIRS})
 target_link_libraries( dd ${MOAB_LIBRARIES} ${EMBREE_LIBRARY} )
@@ -65,25 +68,23 @@ install(EXPORT ddTargets DESTINATION "${DD_CMAKE_CONFIG_LOCATION}")
 export( TARGETS dd FILE "ddTargets.cmake" )
 
 # install headers
-install ( FILES "${CMAKE_CURRENT_SOURCE_DIR}/src/RTI.hpp" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install ( FILES "${CMAKE_CURRENT_SOURCE_DIR}/src/primitives.hpp" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install ( FILES "${CMAKE_CURRENT_SOURCE_DIR}/src/ray.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install ( FILES "${CMAKE_CURRENT_SOURCE_DIR}/src/MOABRay.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install ( FILES "${CMAKE_CURRENT_SOURCE_DIR}/src/MOABDirectAccess.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install ( FILES "${CMAKE_CURRENT_SOURCE_DIR}/src/Vec3da.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install ( FILES "${CMAKE_CURRENT_SOURCE_DIR}/src/constants.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install ( FILES "${CMAKE_CURRENT_SOURCE_DIR}/src/sys.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install ( FILES "${CMAKE_CURRENT_SOURCE_DIR}/src/Vec3fa.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install ( FILES "${CMAKE_CURRENT_SOURCE_DIR}/src/Vec3ba.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install ( FILES "${CMAKE_CURRENT_SOURCE_DIR}/src/Vec3.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-install ( FILES "${CMAKE_CURRENT_SOURCE_DIR}/src/AABB.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
-
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/RTI.hpp" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/primitives.hpp" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/ray.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/MOABRay.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/MOABDirectAccess.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/Vec3da.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/constants.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/sys.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/Vec3fa.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/Vec3ba.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/Vec3.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
+install(FILES "${CMAKE_CURRENT_SOURCE_DIR}/include/double-down/AABB.h" DESTINATION "${DD_INCLUDE_INSTALL_LOCATION}")
 
 # Tests
 enable_testing()
 
 configure_file("${CMAKE_CURRENT_SOURCE_DIR}/cmake/ddConfig.cmake.in" "${DD_CMAKE_CONFIG_LOCATION}/ddConfig.cmake")
-
 
 add_subdirectory(tools)
 add_subdirectory(test)
