+++ include/double-down/Vec3ba.h
