+++ src/RTI.cpp
@@ -6,8 +6,8 @@
 #include "moab/GeomTopoTool.hpp"
 
 // Local
-#include "RTI.hpp"
-#include "AABB.h"
+#include "double-down/RTI.hpp"
+#include "double-down/AABB.h"
 
 
 void error(void* dum, RTCError code, const char* str) {
