+++ test/test_closest.cpp
@@ -2,8 +2,9 @@
 
 #include <iostream>
 
-#include "RTI.hpp"
 #include "test_utils.hpp"
+
+#include "double-down/RTI.hpp"
 
 
 int main() {
