+++ test/test_mb.cpp
@@ -2,7 +2,8 @@
 
 #include <iostream>
 
-#include "RTI.hpp"
+#include "double-down/RTI.hpp"
+
 #include "test_utils.hpp"
 
 int main() {
