+++ test/test_rf.cpp
@@ -2,9 +2,9 @@
 
 #include <iostream>
 
-#include "RTI.hpp"
 #include "test_utils.hpp"
 
+#include "double-down/RTI.hpp"
 
 int main() {
